Update to discord.js@12 Add some custom recode commands

+++ src/commands/admin/pingUser.ts
@@ -1,5 +1,5 @@
-import { Permissions } from "discord.js";
-import { AddCommandMethod, Bot, Context, ICommandSet, Injector, isCreator, isTextChannelContext } from "../../lib";
+import { Permissions, TextChannel } from "discord.js";
+import { AddCommandMethod, Context, ICommandSet, isCreator, isTextChannelContext } from "../../lib";
 import { hasPerm } from "./common";
 
 const enabledPingUser: Set<string> = new Set();
@@ -57,14 +57,24 @@ async function pingUser(context: Context) {
 	const userString = context.args[0];
 	if (context.mentions.users.size === 0 && !userString) {
 		context.reply("Yeah, you're just being stupid now");
+		return;
 	}
 
-	const u = context.mentions.users.size === 0
-		? await Injector.get(Bot).client.fetchUser(userString)
+	const user = context.mentions.users.size === 0
+		? (context.channel as TextChannel).members.find((m) =>
+			m.displayName.toLowerCase() === userString.toLowerCase()
+			|| m.nickname?.toLowerCase() === userString.toLowerCase()
+			|| m.user.username.toLowerCase() === userString.toLowerCase(),
+		)?.user
 		: context.mentions.users.first();
 
+	if (user == null) {
+		context.reply(`Wait... is ${userString} event a person?`);
+		return;
+	}
+
 	await Promise.all([
-		context.sendToChannel(`${u}`),
+		context.sendToChannel(`${user}`),
 		context.delete(),
 	]);
 }
